% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{allFreq}
\alias{allFreq}
\title{Estimate the allele frequencies}
\usage{
allFreq(genotype)
}
\arguments{
\item{genotype}{A matrix of genotypes (n*p)
n = number of individuals
p = number of markers (coded as "All1/All2", ex: "A/A" or "NA/NA" for missing genotype)}
}
\value{
allele frequencies
}
\description{
This function estimates allele frequencies
}
\examples{
data <- as.matrix(read.table(file = "example_genotype_APIS.txt", row.names = 1))
off.genotype <- data[which(rownames(data) \%in\% c(21:1020)),]
freq <- allFreq(off.genotype)
}
\keyword{allele}
\keyword{frequencies}
