% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apis-package.R
\docType{package}
\name{apis-package}
\alias{apis}
\alias{apis-package}
\title{apis: Wrappers to API interfaces}
\description{
Various API interfaces (Druid, ARGOS-CLS, WoK, etc).
}
\author{
\strong{Maintainer}: Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})

Authors:
\itemize{
  \item Mihai Valcu \email{mvalcu@gwdg.de} (\href{https://orcid.org/0000-0002-6907-7802}{ORCID})
}

}
\keyword{internal}
