% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{APIS}
\alias{APIS}
\title{APIS function that assigns with observed data}
\usage{
APIS(off.genotype, sire.genotype, dam.genotype, error = NULL,
  exclusion.threshold = ncol(off.genotype), preselect.Parent = F)
}
\arguments{
\item{off.genotype}{Offspring genotypes | Matrix (n*p)
where n = number of individuals
p = number of markers
rownames(offspring) = labels of offspring
marker coding = "All1/All2" example: "A/A", "A/B", "NA/NA" (for missing genotype)}

\item{sire.genotype}{Sire genotypes | Matrix (n*p)
where n = number of individuals
p = number of markers
rownames(sire) = labels of sires
marker coding = "All1/All2" example: "A/A", "A/B", "NA/NA" (for missing genotype)}

\item{dam.genotype}{Dam genotypes | Matrix (n*p)
where n = number of individuals
p = number of markers
rownames(dam) = labels of dams
marker coding = "All1/All2" example: "A/A", "A/B", "NA/NA" (for missing genotype)}

\item{error}{(default: NULL) The assignment error rate accepted by the user}

\item{exclusion.threshold}{(default: ncol(off.genotype)) Threshold for exclusion (number of mismatches allowed)}

\item{preselect.Parent}{(default: FALSE) Preselection of parents. Can be FALSE, an integer or a vector of two integers (number of sires, numbers of dams)}
}
\value{
pedigree

a log file
}
\description{
This function performs the APIS procedure
}
\examples{
data("genotype_APIS")

result <- APIS(off.genotype = off_full,
               sire.genotype = sire_full,
               dam.genotype = dam_full,
               error = 0.05)
}
\keyword{APIS}
\keyword{assignment}
