% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_dam.R
\docType{data}
\name{APIS_dam}
\alias{APIS_dam}
\title{Example dam genotypes}
\format{
A matrix with 14 rows (one row = one dam) and 100 columns (one column = one marker)
}
\usage{
APIS_dam
}
\description{
Example dam genotypes
}
\keyword{datasets}
