% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_offspring3n.R
\docType{data}
\name{APIS_offspring3n}
\alias{APIS_offspring3n}
\title{Example offspring 3n genotypes}
\format{
A matrix with 50 rows (one row = one offspring) and 100 columns (one column = one marker)
}
\usage{
APIS_offspring3n
}
\description{
Example offspring 3n genotypes
}
\keyword{datasets}
