% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_sire.R
\docType{data}
\name{APIS_sire}
\alias{APIS_sire}
\title{Example sire genotypes from Griot et al. (2019)}
\format{A matrix with 39 rows (one row = one sire) and 100 columns (1 column = one marker)}
\usage{
APIS_sire
}
\description{
Example sire genotypes from Griot et al. (2019)
}
\keyword{datasets}
