% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos.R
\name{argos_devlist}
\alias{argos_devlist}
\title{Argos Device List}
\usage{
argos_devlist(login)
}
\arguments{
\item{login}{A login object obtained from the \code{argos_login} function.}
}
\value{
A data.table containing the list of devices.
}
\description{
This function retrieves a list of devices from the Argos API.
}
\examples{
\dontrun{
crd = config::get(config = "argos_api")
login = argos_login(un = crd$un, pwd = crd$pwd, wsdl_server = crd$wsdl_server)
x = argos_devlist(login)
}
}
