% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos.R
\name{argos_login}
\alias{argos_login}
\title{Argos login token}
\usage{
argos_login(un, pwd, wsdl_server)
}
\arguments{
\item{un}{Username to authenticate with the server.}

\item{pwd}{Password associated with the username.}

\item{wsdl_server}{One or more WSDL server addresses. If multiple addresses are provided, they will be tried in the specified order.}
}
\value{
a character vector.
}
\description{
Argos login token
}
\note{
For security, credentials should either be stored in a configuration file or entered interactively at runtime.
Avoid hardcoding credentials directly in your scripts. See examples for more details.
}
\examples{
\dontrun{
crd = config::get(config = "argos_api")
login = argos_login(un = crd$un, pwd = crd$pwd, wsdl_server = crd$wsdl_server)
}
}
