% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_assignment_power.R
\name{assignment_power}
\alias{assignment_power}
\title{Assignment power}
\usage{
assignment_power(
  sire_genotype,
  dam_genotype,
  ploidy_level = 2,
  verbose = FALSE
)
}
\arguments{
\item{sire_genotype}{matrix of the sire genotypes}

\item{dam_genotype}{matrix of the dam genotypes}

\item{ploidy_level}{ploidy level of the parents}

\item{verbose}{verbose}
}
\value{
the theoretical assignment power calculated with the formula proposed in Vandeputte (2012)
}
\description{
Assignment power
}
\examples{
data("APIS_sire")
data("APIS_dam")

P = assignment_power(sire_genotype = APIS_sire, dam_genotype = APIS_dam)

}
