% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{assignmentPower}
\alias{assignmentPower}
\title{calculte the theoretical assignment power}
\usage{
assignmentPower(sire, dam)
}
\arguments{
\item{sire}{Sire genotypes | Matrix (n*p) where n = number of individuals, p = number of markers
rownames(sire) = labels of sires
marker coding = "All1/All2" example: "A/A", "A/B", "NA/NA" (for missing genotype)}

\item{dam}{Dam genotypes | Matrix (n*p) where n = number of individuals, p = number of markers
rownames(dam) = labels of dams
marker coding = "All1/All2" example: "A/A", "A/B", "NA/NA" (for missing genotype)}
}
\value{
Theoretical assignment power of the marker set
}
\description{
This function calculates the theoretical assignment power of the marker set
}
\examples{
data("APIS_sire")
data("APIS_dam")
assignmentPower(APIS_sire, APIS_dam)
}
\keyword{assignment}
\keyword{exclusion}
\keyword{power}
