% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_data_all_by_id}
\alias{ecotopia_data_all_by_id}
\title{Download all daily data for a device}
\usage{
ecotopia_data_all_by_id(
  logstring,
  id,
  start_date,
  what = "gps",
  verbose = TRUE
)
}
\arguments{
\item{logstring}{login string as returned by \link{ecotopia_login}}

\item{id}{device ID (e.g. 640dab1c6f2d20ea33538465)}

\item{start_date}{first calendar date to retrieve}

\item{what}{data type: "gps", "odba", "sms" or "env"}

\item{verbose}{print \link[httr2:request]{httr2::request} strings. Default to TRUE.}
}
\value{
a data.table
}
\description{
Calls \code{\link[=ecotopia_data_by_day]{ecotopia_data_by_day()}} for every date from \code{start_date} through
today, including \code{start_date}.
}
\examples{
\dontrun{
crd = config::get(config = "druid_api")
logstring = ecotopia_login(crd$generic$un, crd$generic$pwd, crd$kw1, crd$kw2)
x = ecotopia_data_all_by_id(
  logstring,
  id = "650d43ed44c2dcc1a0dc30a9",
  start_date = Sys.Date() - 2,
  what = "gps"
)
}

}
