% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_data_by_day}
\alias{ecotopia_data_by_day}
\title{Download data for a given tag and calendar day}
\usage{
ecotopia_data_by_day(logstring, id, date, what = "gps", verbose = TRUE)
}
\arguments{
\item{logstring}{login string as returned by \link{ecotopia_login}}

\item{id}{device ID (e.g. 640dab1c6f2d20ea33538465)}

\item{date}{calendar date ("2023-01-01")}

\item{what}{data type: "gps", "odba", "sms" or "env"}

\item{verbose}{print \link[httr2:request]{httr2::request} string. Default to TRUE.}
}
\value{
a data.table
}
\description{
Uses the Ecotopia daily endpoint. An empty response body or HTTP 404 response
returns an empty data.table with zero rows and zero columns.
}
\examples{
\dontrun{
crd = config::get(config = "druid_api")
logstring = ecotopia_login(crd$generic$un, crd$generic$pwd, crd$kw1, crd$kw2)
dl = ecotopia_devlist(logstring)
x = ecotopia_data_by_day(logstring, dl$id[204],  Sys.Date() -1, what = "gps")
}

}
