% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_device_page}
\alias{ecotopia_device_page}
\title{Download one page of Ecotopia devices}
\usage{
ecotopia_device_page(logstring, cursor = NULL, verbose = interactive())
}
\arguments{
\item{logstring}{Login string returned by \code{\link[=ecotopia_login]{ecotopia_login()}}.}

\item{cursor}{Device ID cursor returned on the previous page. Use \code{NULL} for
the first page.}

\item{verbose}{Print the \link[httr2:request]{httr2::request} object. Defaults to
\code{\link[=interactive]{interactive()}}.}
}
\value{
A data.table containing one page of devices.
}
\description{
Download one page of Ecotopia devices
}
