% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_gps}
\alias{ecotopia_gps}
\title{Download gps data of a given id after a given datetime}
\usage{
ecotopia_gps(logstring, id, datetime = "2000-01-01T00:00:00Z", verbose = TRUE)
}
\arguments{
\item{logstring}{login string as returned by \link{ecotopia_login}}

\item{verbose}{print \link[httr2:request]{httr2::request} string. Default to TRUE.}

\item{cleanup}{when FALSE (default) returns a cleaned up data.table removing all unnecessary columns.}
}
\value{
a data.table
}
\description{
Download gps data of a given id after a given datetime
}
\examples{
crd = config::get(config = "druid_api")
logstring <- ecotopia_login(crd$generic$un, crd$generic$pwd, crd$kw1, crd$kw2)
dl = ecotopia_devlist(logstring)
x =  ecotopia_gps(logstring, dl$id[100])

}
