% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_login}
\alias{ecotopia_login}
\title{Ecotopia login token ----}
\usage{
ecotopia_login(un, pwd, kw1, kw2, verbose = interactive())
}
\arguments{
\item{un}{Username.}

\item{pwd}{Password.}

\item{kw1}{First password-signature keyword.}

\item{kw2}{Second password-signature keyword.}

\item{verbose}{Print the \link[httr2:request]{httr2::request} object. Defaults to
\code{\link[=interactive]{interactive()}}.}
}
\value{
A character authentication token.
}
\description{
Ecotopia login token ----
}
\note{
Store credentials in a configuration file or enter them
interactively. Do not hardcode them in scripts.
}
\examples{
\dontrun{
crd <- config::get(config = "druid_api")
logstring <- ecotopia_login(
  crd$generic$un,
  crd$generic$pwd,
  crd$kw1,
  crd$kw2
)
}
}
