% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{from_timestamp}
\alias{from_timestamp}
\alias{to_timestamp}
\title{Convert Between ISO 8601 Timestamp Strings and POSIXct Objects}
\usage{
from_timestamp(x)

to_timestamp(x)
}
\arguments{
\item{x}{For \code{from_timestamp()}, a character string representing a timestamp in ISO 8601 format
(e.g., "2024-12-07T06:01:00Z"). For \code{to_timestamp()}, a POSIXct object.}
}
\value{
\itemize{
\item \code{from_timestamp()}: A POSIXct object representing the given timestamp in the UTC timezone.
\item \code{to_timestamp()}: A character string in ISO 8601 format representing the given POSIXct object.
}
}
\description{
These functions facilitate the conversion between ISO 8601 timestamp strings and POSIXct objects.
}
\details{
\itemize{
\item \code{from_timestamp()} converts an ISO 8601 timestamp string to a POSIXct object in the UTC timezone.
\item \code{to_timestamp()} converts a POSIXct object to an ISO 8601 timestamp string in UTC format.
}
}
\examples{
# Convert a timestamp string to a POSIXct object
posix_time <- from_timestamp("2024-12-07T06:01:00Z")

# Convert a POSIXct object back to an ISO 8601 timestamp string
iso_time <- to_timestamp(posix_time)

# Check if conversion and reverse conversion match the original string
identical("2024-12-07T06:01:00Z", from_timestamp("2024-12-07T06:01:00Z") |> to_timestamp())

}
