% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_import.R
\name{import_from_ped}
\alias{import_from_ped}
\title{Import from Plink .ped}
\usage{
import_from_ped(
  ped_file,
  no_fid = FALSE,
  no_parents = FALSE,
  no_sex = FALSE,
  no_pheno = FALSE,
  marker_names = NULL
)
}
\arguments{
\item{ped_file}{name of the ped file (from Plink)}

\item{no_fid}{if "no_fid" parameter was used in plink (default : FALSE)}

\item{no_parents}{if "no_parents" parameter was used in plink (default : FALSE)}

\item{no_sex}{if "no_sex" parameter was used in plink (default : FALSE)}

\item{no_pheno}{if "no_pheno" parameter was used in plink (default : FALSE)}

\item{marker_names}{list of marker names (default : NULL)}
}
\value{
matrix of genotypes for APIS
}
\description{
Import from Plink .ped
}
