% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_import.R
\name{import_from_vcf}
\alias{import_from_vcf}
\title{Import from .vcf}
\usage{
import_from_vcf(vcf_file)
}
\arguments{
\item{vcf_file}{name of the vcf file}
}
\value{
matrix of genotypes for APIS
}
\description{
Import from .vcf
}
