% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineis.R
\name{kineis_retrieve_realtime}
\alias{kineis_retrieve_realtime}
\title{Retrieve KINEIS realtime telemetry from a checkpoint}
\usage{
kineis_retrieve_realtime(
  token,
  api_telemetry_url,
  checkpoint = 0,
  deviceRefs = character(),
  retrieveMetadata = TRUE,
  retrieveRawData = TRUE,
  retrieveDoppler = TRUE,
  retrieveGpsLoc = TRUE,
  retrieveSensors = TRUE,
  retrieveAdditionnalProperties = TRUE,
  datetimeFormat = "DATETIME",
  verbose = interactive()
)
}
\arguments{
\item{token}{either a character access token, or the full list returned by \link{kineis_login}}

\item{api_telemetry_url}{API base url (usually from config)}

\item{checkpoint}{initial checkpoint (0 typically means "last ~6 hours" per KINEIS example)}

\item{deviceRefs}{character vector of device refs; empty means "all available"}

\item{retrieveMetadata, retrieveRawData, retrieveDoppler, retrieveGpsLoc, retrieveSensors, retrieveAdditionnalProperties}{flags}

\item{datetimeFormat}{"DATETIME" unless you know other accepted values}

\item{verbose}{print \link[httr2:request]{httr2::request} string. Defaults to \code{\link[=interactive]{interactive()}}.}
}
\value{
a list with \code{checkpoint}, \code{data} (data.table, possibly empty), and \code{raw} (full parsed JSON)
}
\description{
Retrieve KINEIS realtime telemetry from a checkpoint
}
\examples{
\dontrun{
crd = config::get(config = "kineis_api")
tok = kineis_login(un = crd$un, pwd = crd$pwd, auth_url = crd$auth_url)
out = kineis_retrieve_realtime(tok, api_telemetry_url = crd$api_telemetry_url)
out = kineis_retrieve_realtime(tok, api_telemetry_url = crd$api_telemetry_url,
       checkpoint = attributes(out)$checkpoint )
}
}
