% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineis.R
\name{kineis_save_real_time}
\alias{kineis_save_real_time}
\title{Save KINEIS realtime telemetry to disk (incremental)}
\usage{
kineis_save_real_time(
  path = "~/kineis_local_db.sqlite",
  un,
  pwd,
  auth_url,
  api_telemetry_url
)
}
\description{
Save KINEIS realtime telemetry to disk (incremental)
}
\examples{
\dontrun{
crd = config::get(config = "kineis_api")

kineis_save_real_time(
  un = crd$un,
  pwd = crd$pwd,
  auth_url = crd$auth_url,
  api_telemetry_url = crd$api_telemetry_url
)

}
}
