% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App_APIS_Shiny.R
\name{launch_APIShiny}
\alias{launch_APIShiny}
\title{Shiny App for interactive session of APIS}
\usage{
launch_APIShiny()
}
\value{
void : most results are automatically saved
}
\description{
Launch the shiny interface to use APIS interactively
}
