% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_delta}
\alias{plot_delta}
\title{Plot deltas}
\usage{
plot_delta(log_file, threshold = NULL, simulated_individuals = NULL)
}
\arguments{
\item{log_file}{log file from the APIS_2n() or APIS_3n function}

\item{threshold}{threshold}

\item{simulated_individuals}{names of the simulated individuals}
}
\value{
plot of the distribution of delta
}
\description{
Plot deltas
}
