% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_plot.R
\name{plot_mismatches}
\alias{plot_mismatches}
\title{Plot mismatches}
\usage{
plot_mismatches(log_file, threshold = NULL, simulated_individuals = NULL)
}
\arguments{
\item{log_file}{log file from the APIS_2n() or APIS_3n function}

\item{threshold}{threshold}

\item{simulated_individuals}{names of the simulated individuals}
}
\value{
plot of the distribution of mismatches
}
\description{
Plot mismatches
}
