% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{selectParents}
\alias{selectParents}
\title{Select most likely parents for potent parent pairs tests}
\usage{
selectParents(off.genotype, parent.genotype, parent.sex, n.Parent)
}
\arguments{
\item{off.genotype}{genotype of one offspring}

\item{parent.genotype}{genotype matrix of parent genotypes}

\item{parent.sex}{vector of parents sex}

\item{n.Parent}{vector of number of sires and dams to select}
}
\value{
list of potential sires and dams
}
\description{
This function allows the selection of the most likely parents for assignment, reducing computation time
}
\keyword{APIS}
\keyword{assignment}
\keyword{threshold}
