% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{allFreq}
\alias{allFreq}
\title{Estimate the allele frequencies}
\usage{
allFreq(genotype)
}
\arguments{
\item{genotype}{A matrix of genotypes (n*p)
n = number of individuals
p = number of markers (coded as "All1/All2", ex: "A/A" or "NA/NA" for missing genotype)}
}
\value{
allele frequencies
}
\description{
This function estimates allele frequencies
}
\examples{
data("APIS_offspring")
freq <- allFreq(APIS_offspring)
}
\keyword{allele}
\keyword{frequencies}
