% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_dam.R
\docType{data}
\name{APIS_dam}
\alias{APIS_dam}
\title{Example dam genotypes from Griot et al. (2019)}
\format{A matrix with 1068 rows (one row = one dam) and 100 columns (1 column = one marker)}
\usage{
APIS_dam
}
\description{
Example dam genotypes from Griot et al. (2019)
}
\keyword{datasets}
