% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_offspring.R
\docType{data}
\name{APIS_offspring}
\alias{APIS_offspring}
\title{Example offspring genotypes from Griot et al. (2019)}
\format{A matrix with 1068 rows (one row = one offspring) and 100 columns (1 column = one marker)}
\usage{
APIS_offspring
}
\description{
Example offspring genotypes from Griot et al. (2019)
}
\keyword{datasets}
