% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argos.R
\name{argos_data}
\alias{argos_data}
\title{Retrieve Argos API Data for a Specific Platform ID}
\usage{
argos_data(login, platformId, startDate)
}
\arguments{
\item{login}{A list containing authentication credentials and API information, including the username, password, and the WSDL server URL.}

\item{platformId}{A character string representing the ID of the platform for which data is being requested.}

\item{startDate}{observation date begin of search and end of search. Only the 10 last days are available. If endDate is not specified, the current date is assumed.}
}
\value{
A \code{data.table} containing the retrieved platform data. If an error occurs during the data retrieval process, an empty \code{data.table} is returned.
}
\description{
This function sends a SOAP request to the Argos API to fetch data associated with a given platform ID.
The retrieved data includes platform locations, diagnostic messages, sensor readings, and other relevant information.
}
\note{
The function is limited to retrieving data from the last 20 days, as enforced by the Argos system.
}
\examples{
\dontrun{
crd = config::get(config = "argos_api")
login = argos_login(un = crd$un, pwd = crd$pwd, wsdl_server = crd$wsdl_server)
start_date = (with_tz(Sys.time(), tzone = "UTC") - days(1) ) |> to_timestamp()
x = argos_data(login, platformId=266471, startDate =  start_date)
}

}
