% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_login}
\alias{ecotopia_login}
\title{ecotopia login token}
\usage{
ecotopia_login(un, pwd, kw1, kw2, verbose = TRUE)
}
\arguments{
\item{un}{username}

\item{pwd}{password}

\item{verbose}{print \link[httr2:request]{httr2::request} string. Default to TRUE.}
}
\value{
a character string
}
\description{
ecotopia login token
}
\note{
credentials are saved to a \code{config} file. See examples.
}
\examples{
crd = config::get(config = "druid_api")
logstring = ecotopia_login(crd$generic$un, crd$generic$pwd, crd$kw1,  crd$kw2)

}
