% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/druid.R
\name{ecotopia_login}
\alias{ecotopia_login}
\title{ecotopia login token}
\usage{
ecotopia_login(un, pwd, kw1, kw2)
}
\arguments{
\item{un}{username}

\item{pwd}{password}
}
\value{
a character string
}
\description{
ecotopia login token
}
\examples{
crd = config::get(config = "druid_api")
logstr = ecotopia_login(crd$generic$un, crd$generic$pwd, crd$kw1)

}
