% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineis.R
\name{kineis_login}
\alias{kineis_login}
\title{Kineis login token (JWT) ----}
\usage{
kineis_login(
  un,
  pwd,
  auth_url,
  client_id = "api-telemetry",
  verbose = interactive()
)
}
\arguments{
\item{un}{Username.}

\item{pwd}{Password.}

\item{auth_url}{OAuth token endpoint.}

\item{client_id}{OAuth client ID.}

\item{verbose}{Print the \link[httr2:request]{httr2::request} object. Defaults to
\code{\link[=interactive]{interactive()}}.}
}
\value{
A list containing \code{access_token} and the other token fields.
}
\description{
Kineis login token (JWT) ----
}
\note{
Store credentials in a configuration file or enter them
interactively. Do not hardcode them in scripts.
}
\examples{
\dontrun{
crd <- config::get(config = "kineis_api")
token <- kineis_login(
  un = crd$un,
  pwd = crd$pwd,
  auth_url = crd$auth_url
)
}
}
