% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineis.R
\name{kineis_login}
\alias{kineis_login}
\title{KINEIS login token (JWT)}
\usage{
kineis_login(un, pwd, auth_url, client_id = "api-telemetry", verbose = TRUE)
}
\arguments{
\item{un}{username (from config if missing)}

\item{pwd}{password (from config if missing)}

\item{auth_url}{token endpoint (from config if missing)}

\item{client_id}{oauth client id}

\item{verbose}{print \link[httr2:request]{httr2::request} string. Default to TRUE.}
}
\value{
a list with \code{access_token} plus other token fields
}
\description{
KINEIS login token (JWT)
}
