% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kineis.R
\name{kineis_save_real_time}
\alias{kineis_save_real_time}
\title{Save KINEIS realtime telemetry to disk (incremental)}
\usage{
kineis_save_real_time(
  path = "~/kineis_local_db.sqlite",
  un,
  pwd,
  auth_url,
  api_telemetry_url
)
}
\description{
Save KINEIS realtime telemetry to disk (incremental)
}
\examples{
\dontrun{

kineis_save_real_time()

}
}
