% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{personalThreshold}
\alias{personalThreshold}
\title{Establish personal threshold}
\usage{
personalThreshold(APIS.result, method, threshold = NULL)
}
\arguments{
\item{APIS.result}{APIS function output}

\item{method}{the method for the new threshold | 'delta' for deltas,
'Pmendel' for Mendelian porbabilities, 'exclusion' for mismatches}

\item{threshold}{personal threshold | default values are implemented}
}
\value{
new pedigree from the new threshold
}
\description{
This function allows the user to set up his own threshold
}
\examples{
data("genotype_APIS")

result <- APIS(off.genotype = off,
               sire.genotype = sire,
               dam.genotype = dam,
               error = 0.05)

new.result <- personalThreshold(result, method = 'Pmendel')
}
\keyword{APIS}
\keyword{assignment}
\keyword{threshold}
