% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APIS_functions.R
\name{setThreshold}
\alias{setThreshold}
\title{Set the APIS threshold}
\usage{
setThreshold(ped.log, ped.exclu, nb.mrk, error = NULL)
}
\arguments{
\item{ped.log}{log.like for assignment function}

\item{ped.exclu}{log.exclu for assignment function}

\item{nb.mrk}{Number of markers}

\item{error}{(default: NULL) The assignment error rate accepted by the user}
}
\value{
pedigree

log file
}
\description{
This function calculates the threshold for APIS
}
\keyword{assignment}
